requirements.txt
flask
notion-client
gunicorn
requests
openai
python-dotenv
firebase-admin
pytest
google-auth-oauthlib
google-auth-httplib2
google-api-python-client